backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "fastapi[standard]>=0.119.0",
    "psycopg[binary]>=3.2.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
]
